requirements.txt
appium-python-client
selenium
logging
subprocess-run
pandas
undetected-chromedriver
webdriver-manager
kafka-python
geopandas
matplotlib
shapely
contextily
folium